docker_api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=8.3.0
python-multipart>=0.0.5